Implement dynamic width of open-windows switcher
The rofi theme is modified in order to allow changes in the main box
width using the -width flag.
The Python script counts the number of windows active in the session and
launch rofi accordingly with a custom width.

For simplicity, the max number of apps shown in a single box are four.
Future version of the switcher may include multiple rows and wider box
that allows to contain more apps.