Zsh as default shell
Use ZSH as the default shell for Linux.
The main reason for this change is the fact that Vim and conda work
very well together out of the box.

In this commit I added some features and some plugins to mimic some
of the conveniences of the fish shell.
A new custom theme is used `cleanthefish` which is very similar to
the one used in fish shell.
The conda-zsh-completion plugin has been added in order to toggle
autocompletion of conda environments.
Finally, the aliases now are the same as the one defined for fish.

Some other changes have been done. Namely, "~/home/<username>/" has
been substituted with "$HOME" for improving cross-machines
installations and auto-correction is disabled since it only causes
friction in the usual command line workflow (e.g. vim was
continuously asked to be changed with .vim).