Remove oh-my-zsh
Removed oh-my-zsh and port the zsh environment to a more minimal
setup.
All the required packages have been downloaded and sourced one by
one in .zshrc.
Same thing goes for the prompt theme which has been further
modified and adapted to be oh-my-zsh independent.

The .zshrc file has been simplified as well. The original file was
taken from this gits: https://gist.github.com/scottstanfield/fa1085c225069160225d18b1dc16ee1c
and treated as a base.
All the aliases are moved to a new, dedicated file in the
~/.config/zsh/ folder and sourced in the .zshrc file.

The set is more organized, clean and minimal.