Explicitly set environment variables
Environment variables are explicitly set in ~/.zshenv in order to
follow XDG directory guidelines.
All the respective files and directories have been moved accordingly
(mostly in XDG_CONFIG_HOME or XDG_DATA_HOME).

In this sense also .zshrc is not in the `$HOME` directory and is
moved to `ZDOTDIR` directory. The file was also cleaned up, the
`$PATH` variable definition is moved to `.zshenv` and the
initialization of conda is now sourced as a new file.
Some new aliases has been introduced as well.